+++ src/types/rx-change-event.d.ts
@@ -1,11 +1,11 @@
+import { EventBulk } from './rx-storage';
 import { DeepReadonly } from './util';
 
 export type RxChangeEventBase = {
     readonly eventId: string;
     readonly documentId: string;
-    readonly databaseToken: string;
 
-    // optional does not exist on changes to localdocs of the database
+    // optional, does not exist on changes to localdocs of the database
     readonly collectionName?: string;
 
     // true if the event is about a local document, false if not.
@@ -41,3 +41,28 @@ export type RxChangeEventDelete<DocType> = RxChangeEventBase & {
 
 // TODO remove =any
 export type RxChangeEvent<DocType = any> = RxChangeEventInsert<DocType> | RxChangeEventUpdate<DocType> | RxChangeEventDelete<DocType>;
+
+/**
+ * Internally, all events are processed via bulks
+ * to save performance when sending them over a transport layer
+ * or de-duplicating them.
+ */
+export type RxChangeEventBulk<DocType = any> = EventBulk<RxChangeEvent<DocType>> & {
+    // optional, not given for changes to local documents of a RxDatabase.
+    collectionName?: string;
+    /**
+     * Token of the database instance that created the events.
+     * Used to determine if the events came from another instance over the BroadcastChannel.
+     */
+    databaseToken: string;
+    /**
+     * The storageToken of the RxDatabase that created the events.
+     * Used to ensure we do not process events of other RxDatabases.
+     */
+    storageToken: string;
+    /**
+     * If true, the events belong to some internal stuff like from plugins.
+     * Internal events are not emitted to the outside over the .$ Observables.
+     */
+    internal: boolean;
+}