+++ CHANGELOG.md
@@ -8,6 +8,7 @@ BREAKING:
   - `RxStorageInterface`:
     - Replaced all `Map` with plain json objects so that they can be `JSON.stringify`-ed
     - Replaced typings of event stream to use `EventBulk` and process events in bulks to save performance.
+  - Internally all events are handles via bulks, this saves performance when events are transfered over a WebWorker or a BroadcastChannel.
 
 Bugfixes:
   - Do not throw an error when database is destroyed while a GraphQL replication is running.